backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
apscheduler>=3.10.4
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0